move requires_xyz to common_distributed.py
